Extract reusable function to CoreDataHelper file
