+++ FeedLoader/FeedCache/Infrastructure/Coredata/CoreDataHelpers.swift
@@ -1,0 +1,42 @@
+//
+//  CoreDataHelpers.swift
+//  FeedLoader
+//
+//  Created by AmritPandey on 22/06/22.
+//
+
+import Foundation
+import CoreData
+
+extension NSPersistentContainer {
+    
+    enum LoadingError: Error {
+        case modelNotFound
+        case failedToLoadPersistenceStore(Error)
+    }
+    
+    static func load(modelName name: String, url: URL, in bundle: Bundle) throws -> NSPersistentContainer {
+        guard let model = NSManagedObjectModel.with(name: name, in: bundle) else {
+            throw LoadingError.modelNotFound
+        }
+        
+        let description = NSPersistentStoreDescription(url: url)
+        let container = NSPersistentContainer(name: name, managedObjectModel: model)
+        container.persistentStoreDescriptions = [description]
+        
+        var loadError: Error?
+        container.loadPersistentStores { loadError = $1 }
+        try loadError.map { throw LoadingError.failedToLoadPersistenceStore($0) }
+        return container
+    }
+}
+
+extension NSManagedObjectModel {
+    
+    static func with(name: String, in bundle: Bundle) -> NSManagedObjectModel? {
+        return bundle
+            .url(forResource: name, withExtension: "momd")
+            .flatMap{ NSManagedObjectModel(contentsOf: $0) }
+    }
+}
+
